requirements.txt
aiofiles==24.1.0
chromadb==0.5.5
faiss==1.5.3
fastapi==0.115.0
langchain==0.3.0
langchain_chroma==0.1.4
langchain_community==0.3.0
langchain_core==0.3.1
langchain_huggingface==0.1.0
langchain_openai==0.2.0
langsmith==0.1.121
matplotlib==3.9.2
numpy==2.1.1
openai==1.46.0
pandas==2.2.2
pydantic==2.9.2
rank_bm25==0.2.2
tiktoken==0.7.0
tqdm==4.66.5
uvicorn==0.30.6
python-dotenv==1.0.1